Add backend & frontend to show breadcrumb
one GET route that accepts the path with hyphen separated format and returns the children of that path, in case a file is requested an empty array will return. Context api was used to have the current location in both our components and interact with it accordingly. the location state has a stack like behavior to manage the operations needed. A very basic styling with material ui is in place